fixed the counter test case

+++ src/Counter.sol
@@ -1,8 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.18;
-
 contract Counter {
-    int public counter = 0;
+    int256 public counter = 0;
 
     function inc() public {
         counter++;
